apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.11
pydantic-settings==2.10.1
PyJWT==2.10.1
passlib==1.7.4
